scott schleiser patch:
I’ve been trying out the debugger on 3.1 since Aleks checked in some fixes.  I found a small problem, if an expression is added as a watch, once the expression is no longer in scope every step, or breakpoint causes several errors to occur.  This doesn’t seem to have been a problem under 3.0.



I tracked it down to PyVariable.getVariables().  This method was returning null, but it should return an empty IVariable[].  I have attached a patch to fix the issue.


git-svn-id: svn://localhost/trunk@555 7f4d9e04-a92a-ab41-bea9-970b690ef4a7