Add Debugging features to Interactive Console
Add a new option to the interactive console to launch a debug connection
allowing viewing/editing of variables and expressions view.
This also means (when option is enabled) that the interactive console
appears in the Debug (aka Launch) view and it contains a fake/dummy stack
frame to contain the variables.

The debugging is provided by the existing pydevd + the corresponding
jave classes. However, the transmission to pydevd is via the XML-RPC
(see postCommand in PyDebugTargetConsole). The return path is via the
normal socket.