Add an extension point so that third parties can contribute settings when new interpreters are configured.
This is useful when creating eclipse products and tools that use PyDev
to provide the rich python experience, but need to customize, at a product
level, how the interpreter is configured.

For example, the product may allow an end user to use their own
python install, but wants to add an additional python library that is
shipped as part of the product, perhaps in an osgi bundle.