Fix bug 866189: prevent Unittest errors caused by setUpClass/setUpModule exceptions
When an exception is present in setUpClass or setUpModule methods for Unittest
modules, it causes an error in PyDev's PyUnit interface. The error occurs because
the file path is a property of the test variable only when running a custom test
module, not a class or module set up. Correct this error by obtaining file path
from the error output string instead.

An exception in setUpClass/Module also causes an error when trying to access the
name of the test method being called. Correct this error by creating special cases
for these setup methods.

This commit includes new test cases for these scenarios
(org.python.pydev/pysrc/tests_runfiles/samples/simpleClass_test.py & simpleModule_test.py).