Allow faster file-finding when searching for resource declarations in linked files.
Ctrl+clicking on file resources to find their declarations now performs much faster on
linked files, as doing so no longer prompts the user to select the correct file to open out
of a list of all possible locations for the declaration (which occurs if multiple instances
of the linked file are in the workspace). Instead, the declaration-containing file opened is
the one within the same project as the file currently being worked on.