New extension point pydev_interpreter_provider
Add a new extension point to allow third parties to contribute to finding,
and possibly live installing, python (and jython/ipy) installations in
new ways.

This splits the pre-existing code for locating python/jython/ipy on the
path or in the win32 registry out of the AbstractInterpreterEditor class
hierarchy, making it potentially possible to reuse the code without the
Preferences UI being loaded.