New PyDev Project UI notifies the user if project location contains existing files.
If the project directory is set as a location that includes Python files, a message
appears at the top of the wizard window to notify the user of this. A special message
appears if the location contains a file named "__init__.py".

Also, non-default project locations that overlap with the workspace are prohibited.