Streamline auto-configuration of PyDev interpreters.
1) Auto-configuration of interpreters no longer requires a background PreferenceDialog to
work. Therefore, the dialogs that appear when a project's interpreter is not found
(PyDialogHelpers) don't open a Preferences window if the Auto-Config button is clicked.

2) In the group displaying a project's Python settings (PyProjectPythonDetails), when a
project's interpreter is not set, the hyperlink for configuring an interpreter now opens
a dialog asking if an interpreter should be set manually or automatically. Only manual
config opens a PreferenceDialog.

3) Add a "counter" to PyDialogHelpers for preventing the "configure interpreter" dialog
from appearing while other dialogs are open, particularly those that are also for
interpreter configuration.