Allow the auto-configurer to ignore unfit interpreters, and automate the folder selection step.
1) It is possible for the auto-configurer to find an interpreter that is incompatible with
PyDev before finding one that is compatible, thus causing the auto-interpreter to fail when
a usable interpreter is available. Change the auto-configurer to test every interpreter it
finds until one succeeds or until they all fail, rather than to select the first one it finds.

2) For the auto-interpreter only, automate the step where the user must select the folders
to be added to the SYSTEM pythonpath. The folders that are chosen are the same ones that come
pre-checked in the folder selection dialog for manual configuration.