Rewrite of IPython wrapper to use/extend IPython's InteractiveShell
The original interface to IPython was based on IPython.testing.globalipapp,
this new version is a complete re-write, but it delivers the same
functionality.

The reason for this re-write is that further extensions to enable more
IPython features (such as magics and some hooks) is easier to do with
a subclass of InteractiveShell.

There are many todos marked in this code for items that can/should be
improved to enable better IPython integration. These will be part of
future commits.