In IPython >= 0.11 use more of IPython's completions
Avoid recalculating the same completion matches that PyDev does better
by making a Completer that differs from the default and excludes
the duplicate matcher code.

On all completions, both PyDev and IPython are queried for results
and the results are merged, giving preference to PyDev on duplicate
results.

Creates new IToken types for IPython completions, in particular % magics
now have a star next to them.