Fix some improper labeling of the Breakpoint Properties menu.
The Breakpoint Properties menu has incomplete-looking titles ("Properties for " and "common"),
so replace them with more appropriate ones.

NOTE: The new contents of the labels are constants, and are thus not reactive to the type of
breakpoint being edited (like Java's Breakpoint Properties menu) because of incompatibilities
between PyBreakpoint and Util.getAdapter.