The default selection of interpreter PATH folders matches "Select All not in Workspace".
The folders that are automatically selected to be in the SYSTEM pythonpath when configuring
an interpreter are now the same folders that are chosen when clicking on the "Select All
not in Workspace" button in the list dialog. This is now also the selection of folders made
by the auto-configurer.

In addition, this patch renames some occurrences of "Iron Python" to "IronPython."