Add an "Advanced Auto-Config" option to interpreter configuration.
This config option searches for all valid interpreters automatically, but asks the user to
choose one of them (rather than automatically selecting the first one that works). The
folders to add to the SYSTEM pythonpath are also up to the user.

The old Auto-Config option has been renamed to "Quick Auto-Config".