Improve various aspects of interpreter auto-configuration.
1. Instead of searching for files that exactly match an expected name,
auto-configurers now search for files that begin with a certain string,
so as to allow discovery of different versions of interpreters (for
example, python2.7, python3)

2. Advanced Auto-Config now tests every interpreter it finds, and only
includes non-failing ones in the displayed list of interpreters to
choose from.

3. Except for in the Preferences Menu, configuration of an interpreter
(once it has been selected by an auto-configurer) is performed as a
user-level job, allowing it to be performed in the background.