Respect pydev_interpreter_new_custom_entries extension point
It used to be (prior to the new auto-config changes, 9a423ec) that askUserInOutPath was the method for telling if frpmString was being
called from a user configuring a new interpreter or by restoring a previously stored one.

A previous fix in 67cef68 addressed one place where the meaning of
the variable changed. This fix extends that fix to add additional
interpreter settings with Quick-Config is used and hence askUserInOutPath
is false.