Proof of concept for setting local variables from the debug console.
This unit test demonstrates how to use ctypes to set the value of
local variables in a separate frame.  Since f_locals is created on
demand as a view into locals, modifying it directly does not actually
update the locals in the frame.  This module contains two approaches
for setting locals -- one that sets them directly, and one that
serializes a locals dict back into the frame locals.