Utilities for getting exception handler frames.
These utilities demonstrate one mechanism for determining where an
exception will be caught in a stack frame.  Exception handlers are
generated using a consistent bytecode pattern, and by scanning through
the code objects in a stack frame these methods return information
about where an exception could be handled.

This is potentially useful for implementing filters for
break-on-exception, since it is often desirable to filter out all
exceptions that will be caught at a particular location.

Note that this approach works but would probably be more portable if it
was implemented using the ast module, however this approach has the
advantage of also working if only bytecode is available.