Add implementation files for #@DontTrace feature.
There are often methods that are uninteresting and that just get in the
way while debugging, for instance property accessors.  The #@DontTrace
tag allows you to mark a method that is unimportant.  You can still put a
breakpoint in these methods, but by default when stepping the debugger
will skip over these methods.

To suppress tracing a method, add the tag @DontTrace
to a comment either preceding or on the same line as
the method definition

E.g.:
  #@DontTrace
  def test1():
      pass

   ... or ...

  def test2(): #@DontTrace
      pass

This commit includes a diff adding a should_trace_hook to pydevd, and an
implementation of a should_trace_hook that adds DontTrace support.