PYDEV-429: Corrects checkResourcesToBeChanged in MarkOccurrences.
During a MarkOccurrences, the code path to PyRenameRefactoring should
not call the checkResourcesToBeChanged code, or else files will be
inadvertantely marked for validateEdit().

This modifies the code path so it's called only when filling the
change object.  To make the problem less likely in the future, this
also refactors the reference-searching logic into a separate class
called PyReferenceSearcher.  It reroutes uses of PyRenameEntryPoint
that are used for finding references to PyReferenceSearcher.  Finally,
it removes the fillChangeObject argument from
PyRenameEntryPoint.checkFinalConditions because all remaining uses
always passed _true_ to it.

Change-Id: I648162d8a26d7ffcc39acb07bc5f251649a6e615