Make object contributions only show up in python files
Things like the 2to3 or the analysis actions only make sense when working
with an actual python file. However by letting them be visible based on
the project nature these actions also end up in the context menus of any
other types of editors used for files within a python project. That means
an xml file happening to sit in a python project will show these actions as
well as a simple text file.

So instead let the actions only show up for resources that are of the
python content type.