Tighten dependencies between pydev plugins
This ensures that installing one pydev plugin into an Eclipse installation
installs only compatible versions of all its pydev dependencies as well.
Before this change it was possible to select say org.python.pydev.shared_ui
version 3.1.0 for installation and end up with org.python.pydev.shared_core
version 3.9.2 installed instead of the compatible 3.1.0

I've chosen to make the dependency this tight instead of allowing 3.9.x
as that makes the update script a bit easier as it can automatically
calculate the upper-limit number. With a compatibility of up to 3.10 that
wouldn't be the case and one would have to specify that as well.