Support automatic folding of comments when opening a file
With this new preference users can select to automatically collapse
multi-line comments whenever the file is opened. The idea here is to reduce
the visual noise when opening a file so a user can concentrate on the actual
source.

The default for this is off to keep compatibility with older PyDev versions
so users do not suddenly miss their documentation.