Show/hide modifier column
Since modifiers are ignored for the contributed hovers when the combining hover is activated per
the active state mask, hide the modifier column when in combining hover mode.

Add info to table label explaiing that lowest num is highest priority when changing priority value.

Swapped order of modifier and priority column so the columns that are hidden or shown (preempt and
modifier) are in the same location.