Upgrade to Lucene 6.1.0
The tests all succeed. Changes should not be disruptive. 

Only 'supposition' I made is that for Directories and SearchManagers
there is the need to also supply another boolean called
'writeAllDeletes' which I defaulted to 'false' to be on the safe side.
If needed this can be parametrized just as 'applyAllDeletes'.

Signed-off-by: Sopot Cela <scela@redhat.com>