Add problem markers for python files that declare invalid encodings
Refactors the getPythonFileEncoding methods so that we can tell the
difference between "no encoding was declared" and "the declared
encoding is invalid" so that we can use the information during code
analysis to create problem markers in the latter case.

Previously, when Pydev detected an unsupported encoding, it would
just dump a stacktrace into the log and silently continue without
informing the user. With this change, Pydev will inform the user
instead of polluting the log.

Signed-off-by: Mat Booth <mat.booth@redhat.com>