Debundle lucene from the shared_core plugin
Now that pydev uses identical bundles as is in orbit, there seems
to be no downside to consuming lucene directly from orbit as an
OSGi dep instead of having it embedded inside a plugin.

This change makes that switch and also makes the lucene bundles
available in the pydev p2 repo so that users of older Eclipse that
did not ship with lucene 6.1 (e.g. Neon) can still install pydev.

Signed-off-by: Mat Booth <mat.booth@redhat.com>