When running inside a sandbox, allow pydev to detect and use python interpreters that are installed on the sandbox host.
Because the sandbox host filesystem is mounted with a prefix, but
the process is still launched on the host via the sandbox API, the
interpreterInfo.py script may report lib paths that don't exist
inside the sandbox. This is because python doesn't know that it
was invoked by a sandboxed process.

This change corrects the reported lib paths to use the same prefix
as the interpreter executable allowing pydev to correctly reference
all the installed libs from inside the sandbox.

Signed-off-by: Mat Booth <mat.booth@redhat.com>