Extend syntax highlighter to allow colorizing of variables and properties
This adds new syntax highlighting options for variables and property
expression.

The options will end up in "Preferences -> PyDev -> Editor" and are
denoted as "Variable (approx.)" and "Properties (approx.)". They default
to black.

I chose to add the approximation remark because the python syntax
highlighting is only token based and does not use the AST, therefore the
highlighting can only be approximated. The highlighting seems to be
optimized for speed and is called in a blocking fashion on each text
change, it's also only called for the sections that are marked as changed.

The rule objects that are used for the highlighting are kept for as long as the
editor is open, which enables us to look at previous tokens and unchanged
sections.

So while this change has some limitations it already makes code with lots
of variables and properties a lot nicer to read.