Fix renaming of global variables in modules
There are currently two bugs when trying to rename globally defined variables
in modules:

1. renaming from inside the module where the variable is defined only
renames the variable in that module but not in any modules using the
variable
2. renaming the variable in a module where it's used but not defined ends
up not finding any occurrences for renaming in the corresponding dialog

A short sample setup can be found at the end of the message. This problem
occurs regardless of whether the modules are in the same or different
source folders.

This change enables the PyRenameGlobalProcess to also consider uses of
the variable in the form 'mymodule.var1' and thus find not only the
definition but also all uses of the variable by asking the ScopeAnalysis
to not only check the first part but all attribute parts when finding local
occurrences.

Example that fails to rename the variable properly:

mymodule1.py:
  var1 = "Hello World"

main.py:
  import mymodule1

  print("%s!" % mymodule1.var1)