Use the IScopedPreferences API for analysis preferences
A product shipping PyDev plugins may want to override the default values
for certain preference settings. This can be done by specifying the
intended value inside a plugin_customizations.ini file that is loaded as
part of the product. The Eclipse eco-system will ensure that these values
land in the DefaultScope.INSTANCE preference instead of the values set
via the corresponding PreferenceInitializer classes.

In order to ensure those changed defaults are actually used one either has
to use an EclipsePreferenceStore instance that handles lookup through both
the instance and the default scope, explicitly use the
DefaultScope.INSTANCE for determining the default-value-argument or as
a third option use the IEclipsePreferenceService' API for accessing the
preference keys.

The AnalysisPreference class used neither of those, leading to confusing
behavior as the preference pages themselves would reflect the
default-values from the plugin_customizations.ini file (since they use a
PreferenceStore internally) but the behavior when the code read the
preferences would not reflect that.

Since there's already a ScopeedAnalsysisPreference which handles the
insatnce and default scopes and also enables use of user or project
specific configuration files I ported the code to use that type. There's
been a few cases where no project-adaptable object is available and I
decided to just pass null there. In one case it was easy enough to
extend the caller to pass that information forward as a PyEdit instance
was readily available.