Use ScopedPreferences for CodeFoldingSetter
This enables proper overriding of the default value for these preference
items in a plugin_customization.ini file shipped as part of the Eclipse
product that includes PyDev.

It also enables setting these preferences from a yaml file in the users
home directory in addition to the Eclipse preference databases.

I opted for introducing a getBooleanPreference helper function as otherwise
there'd be a lot of duplicated function arguments for the
IScopedPreferences.getBoolean function. This makes the code quite a bit
easier to read IMO. Since the code has no access to anything that can be
adapted to an project-related object I've opted for just passing null
for that.

I would've liked to reuse the static helper functions in
PyScopedPreferences, but those do not really support using testing-specific
IEclipsePreferences instances for running as part of the tests. So I'm
only re-using the get() function to be able to share the yaml-filename
without duplicating any logic.

Finally I made getPreferences private as part of the patch to better
indicate that there's no outside user for this getter.