Pass a project-adaptable in more cases for the changed preference-readings
This enables the use of project-specific configurations for more of the
preferences and ways they're being used. In particular auto-completion
settings will now obey project-preferences when triggered from within an
editor (instead of the console). The same applies to the CodeFoldingSetter,
it will also take project-specific configuration into account now.